feat: create shares
This feature will create shares with defined owner, consumers and
grant read (usage) privileges on databases and schemas.

Defining which tables and views the should be shared is not implemented
and is for now considered the role of dbt.

Changing the owner of an already existing share is not implemented
because this requires comparing the state of the share in snowflake.
This adds complexity and should only be implemented if it becomes a
occurring problem.